code refactor to archive_utils file
